requirements.txt
ragas
faiss-cpu
openai
python-dotenv
datasets
pydantic
